+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### [Unreleased][HEAD]
 
+### Added
+- adds transparent appearance style for `calcite-button`
+- adds `iconposition` prop to `calcite-button`
+- updates dark theme style for `calcite-button`
+  
+### Fixed
+
+
 ## [v1.0.0-beta.4] - Aug 19th 2019
 
 ### Added
@@ -26,7 +34,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - date picker page-up and page-down buttons
 - pre-render support for existing components
 
-### Fixes
+### Fixed
 - style updates/dark theme for buttons
 - fixed styling of modals in firefox
 - fixed radio-group styling in Edge