+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - adds scale prop to `calcite-radio-group`
 - updates style of  `calcite-radio-group`
+- adds transparent appearance style for `calcite-button`
+- adds `iconposition` prop to `calcite-button`
+- updates dark theme style for `calcite-button`
   
 ### Fixed
 
@@ -33,7 +36,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - date picker page-up and page-down buttons
 - pre-render support for existing components
 
-### Fixes
+### Fixed
 - style updates/dark theme for buttons
 - fixed styling of modals in firefox
 - fixed radio-group styling in Edge