+++ CHANGELOG.md
@@ -7,12 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### [Unreleased][HEAD]
 
+## [v1.0.0-beta.5] - Aug 21th 2019
+
 ### Added
 - adds scale prop to `calcite-radio-group`
 - updates style of  `calcite-radio-group`
 - adds transparent appearance style for `calcite-button`
 - adds `iconposition` prop to `calcite-button`
 - updates dark theme style for `calcite-button`
+- updates theme for `calcite-tree`
 
 ### Fixed
 - fix width of medium/large modals with narrow contents
@@ -50,6 +53,7 @@ Fix issue with previous release.
 
 First initial beta release.
 
+[v1.0.0-beta.5]: https://github.com/ArcGIS/calcite-components/compare/v1.0.0-beta.4...v1.0.0-beta.5 "v1.0.0-beta.5"
 [v1.0.0-beta.4]: https://github.com/ArcGIS/calcite-components/compare/v1.0.0-beta.3...v1.0.0-beta.4 "v1.0.0-beta.4"
 [v1.0.0-beta.3]: https://github.com/ArcGIS/calcite-components/compare/v1.0.0-beta.2...v1.0.0-beta.3 "v1.0.0-beta.3"
 [v1.0.0-beta.2]: https://github.com/ArcGIS/calcite-components/compare/v1.0.0-beta.1...v1.0.0-beta.2 "v1.0.0-beta.2"